Refactor: 회원가입 정보 수정 컴포넌트 정리
